Minor Update/Structure
!+ includes/cub3d.h
   ^^ Added more prototypes

!+ Makefile
   ^^ New file added to the SRC

!+ src/init.c
   ^^ Removed a print_map() function.
   ^^ Moved functions to init_utils2 for norm.
  ^^ Assign texture, ray and player into game in memory_allocate() function instead of init_structure.

++ src/init_utils2.c
   ^^ New file for created since init.c is full.

The Game compiles normally.
